aiops-end-to-end-platform/services/consumer/requirements.txt
kafka-python
numpy
scikit-learn
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
prometheus_client